created products CRUD improved security on users

+++ src/app/user/user.controller.ts
@@ -1,4 +1,5 @@
-import { Body, Controller, Delete, Get, Param, Post, Put } from '@nestjs/common';
+import { Body, Controller, Delete, ForbiddenException, Get, Param, Post, Put, Req } from '@nestjs/common';
+import { Request } from 'express';
 import UserService from './user.service';
 import { UserDetails, UserRegistrationDetails } from './types/user.controllerTypes';
 import { Authenticated } from '../security/decorators/security.decorators';
@@ -28,13 +29,19 @@ export default class UserController {
 
   @Authenticated({ oneOfRoles: [UserRoles.seller] })
   @Put('/:userId')
-  updateUser(@Param('userId') userId: string, @Body() body: UserDetails) {
+  updateUser(@Param('userId') userId: string, @Body() body: UserDetails, @Req() req: Request) {
+    if (req.user.id !== parseInt(userId)) {
+      throw new ForbiddenException('You don\'t have access to update this user')
+    }
     return this.userService.updateUser(parseInt(userId), body);
   }
 
   @Authenticated({ oneOfRoles: [UserRoles.seller] })
   @Delete('/:userId')
-  deleteUser(@Param('userId') userId: string) {
+  deleteUser(@Param('userId') userId: string, @Req() req: Request) {
+    if (req.user.id !== parseInt(userId)) {
+      throw new ForbiddenException('You don\'t have access to delete this user')
+    }
     return this.userService.deleteUser(parseInt(userId));
   }
 }
